% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xenoResponseMetrics.R
\name{computeAUC}
\alias{computeAUC}
\title{AUC - area under the curve
computes difference between AUC of treated vs control normalized by treatment}
\usage{
computeAUC(treatedTab, contTab)
}
\description{
AUC - area under the curve
computes difference between AUC of treated vs control normalized by treatment
}
