% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{deseq2NormFilter}
\alias{deseq2NormFilter}
\title{deseq2NormFilter
Utilizes the geometric mean to normalize the entire table to generate a normalized
list of RNAseq counts filtered for low abundance
Only does column normalization and returns DESeq2 object}
\usage{
deseq2NormFilter(data.table)
}
\arguments{
\item{data.table}{#' @import BiocManager}
}
\description{
deseq2NormFilter
Utilizes the geometric mean to normalize the entire table to generate a normalized
list of RNAseq counts filtered for low abundance
Only does column normalization and returns DESeq2 object
}
