% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{do_deseq_import}
\alias{do_deseq_import}
\title{do_deseq2 import
Uses tximport functionality to properly re-count salmon based estimates and returns
counts
TODO: update to include more than just counts}
\usage{
do_deseq_import(file)
}
\arguments{
\item{file}{in sf format}
}
\description{
do_deseq2 import
Uses tximport functionality to properly re-count salmon based estimates and returns
counts
TODO: update to include more than just counts
}
