% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagOfGenesEnrichment.R
\name{doGSEA}
\alias{doGSEA}
\title{Old plot using clusterProfiler}
\usage{
doGSEA(
  genes.with.values,
  prot.univ,
  prefix,
  useEns = FALSE,
  pathway.plot.size = 3,
  order.by = "NES",
  clean.names = F,
  width = 11,
  height = 8.5,
  gsea_FDR = 0.05,
  compress_output = FALSE
)
}
\arguments{
\item{genes.with.values}{data frame of gene names and values}

\item{prot.univ}{total proteins}

\item{prefix}{used to create file}
}
\description{
Old plot using clusterProfiler
}
