% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtResponseMetrics.R
\name{drcmod}
\alias{drcmod}
\title{drcmod
Modified dose response curve function}
\usage{
drcmod(dose_resp, fctval, curve_type)
}
\value{
model
}
\description{
drcmod
Modified dose response curve function
}
