% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugBiomarkers.R
\name{drugMutationsCor}
\alias{drugMutationsCor}
\title{drugMutationCor
Evaluates correlation betwen allele depth and various mutational metrics}
\usage{
drugMutationsCor(pat.drug, mutData)
}
\arguments{
\item{pat.drug}{output of `getAllDrugStats`}

\item{mutData}{output of ...}
}
\value{
table of drug, gene, metric, and correlation value
}
\description{
drugMutationCor
Evaluates correlation betwen allele depth and various mutational metrics
}
