% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugBiomarkers.R
\name{drugTranscriptCor}
\alias{drugTranscriptCor}
\title{drugTranscriptCor
Evaluates correlation}
\usage{
drugTranscriptCor(drug.tab, metric, rna.tab)
}
\description{
drugTranscriptCor
Evaluates correlation
}
