% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagOfGenesEnrichment.R
\name{geneIdToSymbolMatrix}
\alias{geneIdToSymbolMatrix}
\title{geneIdToSymbolMatrix
This takes a symbol wtih gene ids and maps them to symbols}
\usage{
geneIdToSymbolMatrix(gene.mat, identifiers)
}
\arguments{
\item{identifiers}{data table of identifiers}
}
\description{
geneIdToSymbolMatrix
This takes a symbol wtih gene ids and maps them to symbols
}
