% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtResponseMetrics.R
\name{generate_DR_plots}
\alias{generate_DR_plots}
\title{generate_DR_plots
We need to fit the data to a curve to plot}
\usage{
generate_DR_plots(res, drugID)
}
\description{
generate_DR_plots
We need to fit the data to a curve to plot
}
