% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incucyteMetrics.R
\name{generate_TR_plots}
\alias{generate_TR_plots}
\title{generate_TR_plots}
\usage{
generate_TR_plots(res, drugID)
}
\arguments{
\item{res}{}

\item{drugID}{}
}
\value{

}
\description{
generate_TR_plots
}
