% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xenoResponseMetrics.R
\name{getAllDrugStats}
\alias{getAllDrugStats}
\title{getAllDrugStats}
\usage{
getAllDrugStats(drug.tab)
}
\arguments{
\item{drugData}{table}
}
\description{
computes all PDX drug growth statistics
}
