% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getAllNF1Expression}
\alias{getAllNF1Expression}
\title{getAllNF1Expression collects all data from NF1 processed data in synapse}
\usage{
getAllNF1Expression(syn)
}
\description{
getAllNF1Expression collects all data from NF1 processed data in synapse
}
