% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getNewSomaticCalls}
\alias{getNewSomaticCalls}
\title{james said this is the format for future data}
\usage{
getNewSomaticCalls(syn, fileid, specimen)
}
\description{
james said this is the format for future data
}
