% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getPdxRNAseqData}
\alias{getPdxRNAseqData}
\title{getPdxRNAseqData gets all rna seq counts for xenografts
#'@export}
\usage{
getPdxRNAseqData(syn)
}
\description{
getPdxRNAseqData gets all rna seq counts for xenografts
#'@export
}
