% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagOfGenesEnrichment.R
\name{hypothesisTestPlotsDE}
\alias{hypothesisTestPlotsDE}
\title{#### Optional function to visualize hypothesis testing}
\usage{
hypothesisTestPlotsDE(dds)
}
\description{
#### Optional function to visualize hypothesis testing
}
