% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{loadPDXData}
\alias{loadPDXData}
\title{loadPDXData gets data from synapse and stores them as global variables}
\usage{
loadPDXData()
}
\description{
loadPDXData gets data from synapse and stores them as global variables
}
