% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mergeMutData}
\alias{mergeMutData}
\title{until we have fully processed new data, grab missing samples from the old data}
\usage{
mergeMutData(mutData, newMutData)
}
\description{
until we have fully processed new data, grab missing samples from the old data
}
