% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{normDiffEx}
\alias{normDiffEx}
\title{normDiffEx
Normalizes differential data using limma and voom}
\usage{
normDiffEx(data.table)
}
\description{
normDiffEx
Normalizes differential data using limma and voom
}
