% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSets.R
\name{plotDrugData}
\alias{plotDrugData}
\title{plotHistogram of drug data}
\usage{
plotDrugData(drugData)
}
\arguments{
\item{drugData}{table of drug data to be plotted}
}
\description{
plotHistogram of drug data
}
