% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagOfGenesEnrichment.R
\name{plotGenesetResults}
\alias{plotGenesetResults}
\title{ploGenesetResults
Internal function to plot results of cluster profile}
\usage{
plotGenesetResults(
  res,
  prefix,
  pathway.plot.size = 3,
  order.by = "NES",
  clean.names = F,
  width = 11,
  height = 8.5
)
}
\description{
ploGenesetResults
Internal function to plot results of cluster profile
}
