% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSets.R
\name{plotGrowthForBatch}
\alias{plotGrowthForBatch}
\title{plotGrowthForBatch
uses Xeva pacage to plot growth of treatment vs. control}
\usage{
plotGrowthForBatch(xeva.obj, batch)
}
\description{
plotGrowthForBatch
uses Xeva pacage to plot growth of treatment vs. control
}
