% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSets.R
\name{plotIncTreatmentByDrug}
\alias{plotIncTreatmentByDrug}
\title{plotIncTreatmentByDrug}
\usage{
plotIncTreatmentByDrug(dt, sample)
}
\arguments{
\item{dt}{mt data table}
}
\description{
plotIncTreatmentByDrug
}
