% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSets.R
\name{plotMTTreatmentByDrug}
\alias{plotMTTreatmentByDrug}
\title{plotMTTreatmentByDrug}
\usage{
plotMTTreatmentByDrug(samps, drugs, syn)
}
\arguments{
\item{samps}{list of sample identifiers}

\item{drugs}{list of drug names}

\item{syn}{synapse instance}
}
\description{
plotMTTreatmentByDrug
}
