% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSets.R
\name{plotMTTreatmentBySample}
\alias{plotMTTreatmentBySample}
\title{plotMTTreatmentBySample}
\usage{
plotMTTreatmentBySample(dt, drug)
}
\arguments{
\item{dt}{mt data table}
}
\description{
plotMTTreatmentBySample
}
