% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagOfGenesEnrichment.R
\name{plotOldGSEA}
\alias{plotOldGSEA}
\title{Old plot using clusterProfiler}
\usage{
plotOldGSEA(genes.with.values, prot.univ, prefix, useEns = FALSE)
}
\arguments{
\item{genes.with.values}{data frame of gene names and values}

\item{prot.univ}{total proteins}

\item{prefix}{used to create file}
}
\description{
Old plot using clusterProfiler
}
