% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSets.R
\name{plotPDXTreatmentBySample}
\alias{plotPDXTreatmentBySample}
\title{plotPDXTreatmentBySample}
\usage{
plotPDXTreatmentBySample(dt)
}
\arguments{
\item{dt}{is the drugDatatable}
}
\description{
plotPDXTreatmentBySample
}
