% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagOfGenesEnrichment.R
\name{plotTopGenesHeatmap}
\alias{plotTopGenesHeatmap}
\title{plotTopGenesHeatmap}
\usage{
plotTopGenesHeatmap(
  de.out,
  counts,
  identifiers,
  myvar,
  var.ID,
  adjpval = 0.5,
  upload = FALSE,
  path = ".",
  parentID = NULL
)
}
\arguments{
\item{data}{matrix}

\item{txid, }{genename identifieres}

\item{str}{var}

\item{df}{of var}
}
\description{
Filters and plots expression matrix
}
\author{
Jess
}
