% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSets.R
\name{plotTumorGrowthCorrelations}
\alias{plotTumorGrowthCorrelations}
\title{plotTumorGrowthCorrelations}
\usage{
plotTumorGrowthCorrelations(drugGeneCors, minCor = 0.8)
}
\arguments{
\item{drugGeneCors}{the output of `drugMutationsCor`}

\item{minCor}{absolutely correlation to plot}
}
\description{
plotTumorGrowthCorrelations
}
