% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incucyteMetrics.R
\name{process_IncucyteDrugData}
\alias{process_IncucyteDrugData}
\title{process_IncucyteDrugData}
\usage{
process_IncucyteDrugData(
  master.table,
  upload = FALSE,
  path = ".",
  parentID = NULL
)
}
\arguments{
\item{master.table}{}

\item{upload}{}

\item{path}{}

\item{parentID}{}
}
\description{
process_IncucyteDrugData
}
