% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapseUtil.R
\name{queryPartSynapseTable}
\alias{queryPartSynapseTable}
\title{query synapse table more selectively
This is how you get data from the project}
\usage{
queryPartSynapseTable(tableid, query1, query2)
}
\arguments{
\item{tableid}{synapse id of table}

\item{query1, }{query2}
}
\description{
queries synapse table
}
