% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xenoResponseMetrics.R
\name{statsForDrugPatient}
\alias{statsForDrugPatient}
\title{statsForDrugPatient}
\usage{
statsForDrugPatient(indivId, treat, batch)
}
\description{
Function that calculates all stats for each drug/patient combo
}
