% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xevaProcessing.R
\name{tidiedTableToExpressionSet}
\alias{tidiedTableToExpressionSet}
\title{This function takes a tidied table of data and makes into expresion dataset
can be used for RNASeq
TODO: expand for genomic data and latent variables}
\usage{
tidiedTableToExpressionSet(
  tidied.tb,
  featureVar = "totalCounts",
  nonFeatures = c("Symbol", "path", "zScore", "parent")
)
}
\description{
This function takes a tidied table of data and makes into expresion dataset
can be used for RNASeq
TODO: expand for genomic data and latent variables
}
