% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtResponseMetrics.R
\name{TryFit}
\alias{TryFit}
\title{TryFit
Takes the dose response data and fits it to various models}
\usage{
TryFit(
  dose_response,
  fixed = c(NA, NA, NA, NA),
  names = c(NA, NA, NA, NA),
  nan.handle = c("LL4", "L4"),
  curve_type
)
}
\value{
fitted curve
}
\description{
TryFit
Takes the dose response data and fits it to various models
}
