% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dataFromSynTable}
\alias{dataFromSynTable}
\title{gets a list of synapse ids and binds them together}
\usage{
dataFromSynTable(tab, syn, colname)
}
\description{
gets a list of synapse ids and binds them together
}
