% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagOfGenesEnrichment.R
\name{ds2FactorDE}
\alias{ds2FactorDE}
\title{ds2FactorDE}
\usage{
ds2FactorDE(dds, ids1, ids2, name)
}
\arguments{
\item{dds}{DESeq object}

\item{ids1}{Sample ids}

\item{ids2}{Other sample ids}

\item{name}{for condition}
}
\description{
ds2FactorDE
}
\author{
Sara
}
