% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getPdxRNAseqData}
\alias{getPdxRNAseqData}
\title{getPdxRNAseqData gets all rna seq counts for xenografts
#'@export
DEPRACATED}
\usage{
getPdxRNAseqData(syn)
}
\arguments{
\item{syn}{synapse item from}
}
\description{
getPdxRNAseqData gets all rna seq counts for xenografts
#'@export
DEPRACATED
}
