% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{loadMicrotissueMetadata}
\alias{loadMicrotissueMetadata}
\title{loadMicrotissueMetadata loads and stores microtissue metata to global variable}
\usage{
loadMicrotissueMetadata()
}
\description{
loadMicrotissueMetadata loads and stores microtissue metata to global variable
}
