% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{parseSynidListColumn}
\alias{parseSynidListColumn}
\title{parses a list of synapse ids, like form `syn24215021`}
\usage{
parseSynidListColumn(list.column)
}
\arguments{
\item{list.column}{Name of column to select and unlist}
}
\value{
list of lists
}
\description{
parses a list of synapse ids, like form `syn24215021`
}
