% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSets.R
\name{plotMTTreatmentByDrug}
\alias{plotMTTreatmentByDrug}
\title{plotMTTreatmentByDrug}
\usage{
plotMTTreatmentByDrug(dt, sample)
}
\arguments{
\item{dt}{mt data table}
}
\description{
plotMTTreatmentByDrug
}
