% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagOfGenesEnrichment.R
\name{plotTopGenesHeatmap}
\alias{plotTopGenesHeatmap}
\title{plotTopGenesHeatmap}
\usage{
plotTopGenesHeatmap(
  de.out,
  dds,
  identifiers,
  myvar,
  patients = NULL,
  adjpval = 0.5,
  upload = FALSE,
  path = ".",
  parentID = NULL,
  newVar = ""
)
}
\arguments{
\item{de.out}{diffex results}

\item{dds, }{DESEq object}

\item{identifiers}{mapping to gene name}

\item{myvar}{is name of variable}
}
\description{
Filters and plots expression matrix
}
\author{
Jess
}
